topics/apis/data-pagination/scripts/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
python-dotenv==0.14.0
requests==2.24.0
urllib3==1.25.10
